card_reader/requirements.txt
requests==2.19.1
python-dotenv==0.10.1